requirements.txt
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
dnspython==2.6.1
idna==3.6
packaging==24.0
pycparser==2.22
pymongo==4.6.3
pymongocrypt==1.9.1
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.1
